Refactor project structure so the source is easier to find
Change-type: patch
