--- balenasign/app/balenasign/apikey.py
@@ -1,9 +1,0 @@
-from balenasign.config import CONFIG
-
-
-def get(key, required_scopes=None):
-    username = CONFIG.api_keys.get(key, None)
-    if username is None:
-        return None
-
-    return {"uid": username}
