+++ docker-compose.yml
@@ -142,6 +142,6 @@ services:
       io.balena.features.supervisor-api: 1
 
   balenasign:
-    build: balenasign
+    build: .
     privileged: true
     <<: [*base-service]
